Fix namespace and update MainPage to use ViewModel
Updated the `x:Class` attribute in `MainPage.xaml` to correct the namespace.
Added a `using` directive for the view model in `MainPage.xaml.cs`.
Modified the `MainPage` constructor to accept a `MainViewModel` instance.
Removed the click counter functionality and set the `BindingContext` to the provided view model, enabling navigation capabilities.